+++ .gitignore
@@ -1,0 +1,2 @@
+/project_lis3/data/
+/project_lis3/result/
